Various UI/UX Fixes

+++ migrations/utils/writeToConfig.js
@@ -1,10 +1,13 @@
 const fs = require("fs");
 const path = require("path");
 
-const CONFIG_FILE_PATH = path.join(__dirname, "..", "..", "app", "config.json");
-
 const writeToConfig = config => {
-  fs.copyFileSync(CONFIG_FILE_PATH, `${CONFIG_FILE_PATH}.bak`);
+  const configName = `config.${config.network}.json`;
+  const CONFIG_FILE_PATH = path.join(__dirname, "..", "..", "app", configName);
+
+  if (fs.existsSync(CONFIG_FILE_PATH)) {
+    fs.copyFileSync(CONFIG_FILE_PATH, `${CONFIG_FILE_PATH}.bak`);
+  }
   const existingConfig = JSON.parse(fs.readFileSync(CONFIG_FILE_PATH));
   const newConfig = Object.assign({}, existingConfig, config);
   fs.writeFileSync(CONFIG_FILE_PATH, JSON.stringify(newConfig, null, 2));
